fix:字段修复，循环依赖问题

+++ listen-common/src/main/java/com/example/common/dto/TestDto.java
@@ -36,7 +36,7 @@ public class TestDto {
     /**
      * 测试分数
      */
-    private int Score;
+    private Float Score;
 
     private Date testTime;
